Correct info on mixed-precision field collection iterators. (#61)
Field collection iterators had a helper function to enable the user to
iterate over real fields. This didn't work when fields were mixed
precision. I have removed the helper function so the user now has to use
the underlying iterator - which does support mixed precision. The docs
need updating so the comment that describes the helper function not
working for mixed precision needs to be removed.