Document use of the lfric mpi module (#130)
Closes: #113

This adds some documentation for the lfric_mpi object. Some
documentation falls into "how to use" and some into "how it works" - so
I have added docs in both sections. It also includes adding a link to
the code review primer as there wasn't one on trunk - and it was giving
me warnings when I built the docs.

This is not perfect documentation, but I think it's a good start and
something that can be worked on.

---------

Co-authored-by: Mike Hobson <mike.hobson@cazld00001C.spice.sc.metoffice.gov.uk>
Co-authored-by: Andrew Coughtrie <24609575+andrewcoughtrie@users.noreply.github.com>